Abstract types to a dictionary file
