+++ server-ts/types/store.d.ts
@@ -1,0 +1,13 @@
+import gql from 'graphql-tag';
+
+export interface Store {
+  id?: string;
+  coordinates?: {
+    lat: number;
+    lng: number;
+  };
+  address?: string;
+  name?: string;
+  createdAt?: Date;
+  updatedAt?: Date;
+}
